Remove original DefinitionResolver, remove "Tolerant" on file names
