+++ src/LanguageServer.php
@@ -101,7 +101,7 @@ class LanguageServer extends AdvancedJsonRpc\Dispatcher
     protected $projectIndex;
 
     /**
-     * @var TolerantDefinitionResolver
+     * @var DefinitionResolver
      */
     protected $definitionResolver;
 
@@ -187,7 +187,7 @@ class LanguageServer extends AdvancedJsonRpc\Dispatcher
             $this->globalIndex = new GlobalIndex($stubsIndex, $this->projectIndex);
 
             // The DefinitionResolver should look in stubs, the project source and dependencies
-            $this->definitionResolver = new TolerantDefinitionResolver($this->globalIndex);
+            $this->definitionResolver = new DefinitionResolver($this->globalIndex);
 
             $this->documentLoader = new PhpDocumentLoader(
                 $this->contentRetriever,
