+++ tests/Server/TextDocument/ParseErrorsTest.php
@@ -6,7 +6,7 @@ namespace LanguageServer\Tests\Server\TextDocument;
 use PHPUnit\Framework\TestCase;
 use LanguageServer\Tests\MockProtocolStream;
 use LanguageServer\{
-    Server, Client, LanguageClient, ClientHandler, PhpDocumentLoader, TolerantDefinitionResolver
+    Server, Client, LanguageClient, ClientHandler, PhpDocumentLoader, DefinitionResolver
 };
 use LanguageServer\Index\{Index, ProjectIndex, DependenciesIndex};
 use LanguageServer\ContentRetriever\FileSystemContentRetriever;
@@ -40,7 +40,7 @@ class ParseErrorsTest extends TestCase
             }
         };
         $projectIndex = new ProjectIndex(new Index, new DependenciesIndex);
-        $definitionResolver = new TolerantDefinitionResolver($projectIndex);
+        $definitionResolver = new DefinitionResolver($projectIndex);
         $loader = new PhpDocumentLoader(new FileSystemContentRetriever, $projectIndex, $definitionResolver);
         $this->textDocument = new Server\TextDocument($loader, $definitionResolver, $client, $projectIndex);
     }
